got the initialize game to work
