+++ client/src/actions/types.js
@@ -2,6 +2,8 @@ export const GET_ERRORS = "GET_ERRORS";
 export const USER_LOADING = "USER_LOADING";
 export const SET_CURRENT_USER = "SET_CURRENT_USER";
 export const DROP_TILE = 'DROP_TILE';
+export const RESET_GAME = 'RESET_GAME';
+export const SET_WINNER = 'SET_WINNER';
